feat: use context providers for user on both map and login page
